Completion of timeLogg end-point Timelogging creation/putting and getting is now implemented
Inputverification on checking out is now completed. If checkout is attemped. In
which, surpasses 24 hours. Then stop time must be added to the request.
This is to cover all cases where people forget to logg out.

An endpoint is created to make it possible to extract statistics over a
given time. Entries will be delivered with start and end DateTime with
the difference as a seperate field.